Create README.md, commit #14
This last commit summarizes the basis of what the tutorial was like for me. In the last chapter I gained better understanding of design process and the history of fonts. I was able to move 'figcaption' in reference to the image and continue to practice styling and alignment properties with their value.